NXP-31730: Make sure blob exists on write with CachingBlobStore wrapper

+++ modules/core/nuxeo-core-api/src/main/java/org/nuxeo/ecm/core/blob/CachingBlobStore.java
@@ -132,7 +132,7 @@ public class CachingBlobStore extends AbstractBlobStore {
         String key = blobWriteContext.getKey(); // may depend on write observer, for example for digests
 
         // when using deduplication, check if it's in the cache already
-        if (blobWriteContext.useDeDuplication() && getFileFromCache(key, true).isPresent()) {
+        if (blobWriteContext.useDeDuplication() && getFileFromCache(key, true).isPresent() && exists(key)) {
             // delete tmp file, not needed anymore
             cacheStore.deleteBlob(tmpKey);
             return key;
